llm_talktable/requirements.txt
llm
PyYAML